Consolidate duplicate PointcutMutableSeriesTest into borg.trikeshed.mutable
